Refactor the checkpoint saver and engines.
