+++ dlrover/python/tests/test_elastic_training_agent.py
@@ -31,8 +31,8 @@ from dlrover.python.elastic_agent.master_client import (
 )
 from dlrover.python.elastic_agent.monitor.training import TorchTrainingMonitor
 from dlrover.python.elastic_agent.torch.ckpt_saver import (
+    CommonDirCheckpointSaver,
     AsyncCheckpointSaver,
-    CheckpointSaver,
 )
 from dlrover.python.elastic_agent.torch.training import (
     ElasticLaunchConfig,
@@ -283,8 +283,8 @@ class ElasticTrainingAgentRunTest(unittest.TestCase):
             start_method=self.config.start_method,
             log_dir=self.config.log_dir,
         )
-        saver = AsyncCheckpointSaver("/tmp/test")
-        CheckpointSaver._saver_instance = saver
+        saver = CommonDirCheckpointSaver("/tmp/test")
+        AsyncCheckpointSaver._saver_instance = saver
         agent._save_ckpt_to_storage()
         agent._stop_workers_to_restart()
