+++ dlrover/trainer/tests/torch/megatron_ckpt_test.py
@@ -23,7 +23,7 @@ import torch.nn.functional as F
 import torch.optim as optim
 
 from dlrover.python.common.constants import CheckpointConstant
-from dlrover.python.elastic_agent.torch.ckpt_saver import CheckpointSaver
+from dlrover.python.elastic_agent.torch.ckpt_saver import AsyncCheckpointSaver
 from dlrover.trainer.torch.megatron import async_checkpoint
 from dlrover.trainer.torch.megatron.async_checkpoint import (
     MegatronCheckpointManager,
@@ -55,14 +55,14 @@ class SimpleNet(nn.Module):
 
 class MegatrionCheckpointTest(unittest.TestCase):
     def setUp(self):
-        CheckpointSaver._saver_instance = None
-        CheckpointSaver.start_async_saving_ckpt()
+        AsyncCheckpointSaver._saver_instance = None
+        AsyncCheckpointSaver.start_async_saving_ckpt()
         mock_func = mock_get_tracker_filename
         async_checkpoint.get_checkpoint_tracker_filename = mock_func
 
     def tearDown(self) -> None:
-        if CheckpointSaver._saver_instance:
-            CheckpointSaver._saver_instance.close()
+        if AsyncCheckpointSaver._saver_instance:
+            AsyncCheckpointSaver._saver_instance.close()
 
     def test_save_load(self):
         os.environ["LOCAL_RANK"] = "0"
